requirements.txt
Fabric==1.8.0
Flask==0.10.1
Flask-DebugToolbar==0.8.0
Flask-Funnel==0.1.10
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==1.0
Flask-Script==0.6.2
Flask-Security==1.6.9
Flask-WTF==0.9.2
Jinja2==2.7.1
MarkupSafe==0.18
Pillow==2.3.0
SQLAlchemy==0.8.2
WTForms==1.0.5
Werkzeug==0.9.4
argparse==1.2.1
blinker==1.3
boto==2.22.1
distribute==0.6.34
ecdsa==0.10
gunicorn==18.0
itsdangerous==0.23
nose==1.3.0
paramiko==1.12.0
passlib==1.6.1
py==1.4.20
py-bcrypt==0.4
pycrypto==2.6.1
pyelasticsearch==0.6
pyjade==2.0.3
pytest==2.5.2
requests==1.2.3
simplejson==3.3.1
six==1.4.1
wsgiref==0.1.2
